added scaffolding for block_room class. Created Helper_Method class and pulled methods from reservation and booking manager so that these methods stand alone and can be reused elsewhere.
