+++ lib/block_room.rb
@@ -1,0 +1,11 @@
+require 'date'
+require_relative 'room'
+require_relative 'reservation' 
+
+module Hotel
+  class Block_Room
+
+
+
+  end
+end
