+++ lib/booking_manager.rb
@@ -2,6 +2,7 @@ require 'date'
 require 'pry'
 require_relative 'room'
 require_relative 'reservation'
+require_relative 'helper_method'
 
 module Hotel
   class Booking_Manager
@@ -18,8 +19,7 @@ module Hotel
     def reserve_room(input)
 
       search_room_availability(input[:check_in_date], input[:check_out_date])
-      
-      connected_room_number = find_room_number(input[:room_number])
+      connected_room_number = Hotel::Helper_Method.find_room_number(@rooms, input[:room_number])
 
       reserve = { name: input[:name],
         room_number: connected_room_number,
@@ -27,25 +27,17 @@ module Hotel
         check_out_date: input[:check_out_date]
       }
       reservation = Reservation.new(reserve)
-      connect_reservation_to_room(connected_room_number, reservation)
+      Hotel::Helper_Method.connect_reservation_to_room_and_sort(connected_room_number, reservation)
       @hotel_reservations << reservation
-      sort_reservations(@hotel_reservations)
+      Hotel::Helper_Method.sort_reservations(@hotel_reservations)
     end
 
-    def total_cost_of_stay(reservation)
-      return (reservation.nights_of_stay.length * ROOM_COST)
-    end
-
-    def sort_reservations(list_of_reservations)
-      list_of_reservations.sort! { |a,b| a.check_in_date <=> b.check_in_date }
-    end
 
     def search_room_availability(check_in_date, check_out_date)
-      possible_nights_of_stay = Hotel::Reservation.generate_nights(check_in_date, check_out_date)
+      possible_nights_of_stay = Hotel::Helper_Method.generate_nights(check_in_date, check_out_date)
       vacant_rooms = []
-      booked_rooms = []
       @rooms.each do |room|
-        if binary_search_list_of_reservations(room.reservations, possible_nights_of_stay) == false
+        if Hotel::Helper_Method.binary_search_list_of_reservations(room.reservations, possible_nights_of_stay) == false
           vacant_rooms << room.room_number
         end
       end
@@ -63,14 +55,8 @@ module Hotel
       return reservations_on_date = @hotel_reservations.find_all { |reservation| reservation.nights_of_stay == date }
     end
 
-
-    def find_room_number(find_room_number)
-      return @rooms.find { |room| room.room_number == find_room_number }
-    end
-
-    def connect_reservation_to_room(connected_room_number, reservation)
-      connected_room_number.reservations << reservation
-      sort_reservations(connected_room_number.reservations)
+    def total_cost_of_stay(reservation)
+      return (reservation.nights_of_stay.length * ROOM_COST)
     end
 
     def load_rooms
@@ -80,33 +66,6 @@ module Hotel
         rooms << room
       end
       return rooms
-    end
-
-    def binary_search_list_of_reservations(array_of_reservations, array_of_possible_dates)
-
-      min = 0
-      max = array_of_reservations.length
-
-      if array_of_reservations.first == nil
-        return false
-      end
-
-      while min < max
-        mid = (min + max )/ 2
-        if !(array_of_possible_dates & array_of_reservations[mid].nights_of_stay).empty?
-          return true
-        elsif array_of_reservations[mid].check_in_date > array_of_possible_dates.first
-          max = mid - 1
-        elsif array_of_reservations[mid].check_in_date < array_of_possible_dates.first
-          min = mid + 1
-        end
-      end
-
-      if !(array_of_possible_dates & array_of_reservations[0].nights_of_stay).empty?
-        return true
-      end
-
-      return false
     end
 
   end
